+++ src/zck.c
@@ -166,7 +166,7 @@ int main (int argc, char *argv[]) {
 
                     }
                     prev_srpm = next_srpm;
-                    printf("Compressing %li bytes\n", next-found);
+                    printf("Compressing %li bytes\n", (long)(next-found));
                     if(zck_write(zck, found, next-found) < 0)
                         exit(1);
                     if(zck_end_chunk(zck) < 0)
@@ -176,7 +176,8 @@ int main (int argc, char *argv[]) {
                     if(search > data + in_size)
                         search = data + in_size;
                 } else {
-                    printf("Completing %li bytes\n", data+in_size-found);
+                    printf("Completing %li bytes\n",
+                           (long)(data+in_size-found));
                     if(zck_write(zck, found, data+in_size-found) < 0)
                         exit(1);
                     if(zck_end_chunk(zck) < 0)
@@ -210,7 +211,7 @@ int main (int argc, char *argv[]) {
                 } else {
                     cur_loc = data + in_size;
                 }
-                printf("Completing %li bytes\n", cur_loc-start);
+                printf("Completing %li bytes\n", (long)(cur_loc-start));
                 if(zck_write(zck, start, cur_loc-start) < 0)
                     exit(1);
                 if(zck_end_chunk(zck) < 0)
